+++ wear/src/main/java/com/flex/wear/WearApplication.kt
@@ -1,5 +1,0 @@
-package com.flex.wear
-
-import android.app.Application
-
-class WearApplication : Application()